third_party/WebKit/LayoutTests/vr/getPose_samewithinframe.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/fake-vr-displays.js"></script>
<script src="resources/mock-vr-service.js"></script>
<script src="resources/test-constants.js"></script>
<script>
let fakeDisplays = fakeVRDisplays();

vr_test( (t, mock_service) => {
  return navigator.getVRDisplays().then( (displays) => {
    var display = displays[0];
    var expected_pose = VALID_POSE;
    mock_service.mockVRDisplays_[0].setPose(expected_pose);
    var counter = 0;

    function onFrame() {
      var pose1 = display.getPose();
      var pose2 = display.getPose();
      t.step( () => {
        assert_not_equals(pose1, null);
        assert_not_equals(pose2, null);
      }, "getPose returned something");
      t.step( () => {
        let counter = 0;
        for (let field in pose1) {
          counter++;
          assert_equals(pose1[field].length, pose2[field].length);
          assert_greater_than(pose1[field].length, 0);
          for (let i = 0; i < pose1[field].length; i++) {
            assert_equals(pose1[field][i], pose2[field][i]);
          }
        }
        assert_equals(counter, 6);
      }, "poses match");
      t.done();
    }

    display.requestAnimationFrame(onFrame);
  }, (err) => {
    t.step( () => {
      assert_unreached("getVRDisplays rejected");
    });
  });
}, [fakeDisplays["Pixel"]],
"getPose returns the same data within a single frame");

</script>
